feat: remove categorySlug routing path
