+++ app/g/[groupSlug]/market/demands/page.tsx
@@ -1,15 +1,28 @@
 'use client';
 
-import { redirect } from 'next/navigation';
 import { groupVar } from '@/lib/apollo/cache';
 import { useReactiveVar } from '@apollo/client';
+import DemandFeed from '@/components/demands/demand-feed';
+import { useDealStatus } from '@/hooks/use-deal-status';
+import { useSearchParams } from 'next/navigation';
 
-function Page() {
-  const group = useReactiveVar(groupVar);
-  if (!group) return <div />;
-
-  const defaultCategorySlug = group?.productCategories[0].slug;
-  return redirect(`demands/${defaultCategorySlug}`);
+export interface DemandsPageProps {
+  params: {
+    groupSlug: string;
+  };
 }
 
-export default Page;
+function DemandsPage({ params: { groupSlug } }: DemandsPageProps) {
+  const group = useReactiveVar(groupVar);
+  const searchParams = useSearchParams();
+  const categorySlug = searchParams.get('category');
+  const category = group?.productCategories.find(
+    (c) => c.slug === categorySlug,
+  );
+  const status = useDealStatus();
+
+  if (!group) return <div />;
+  return <DemandFeed categoryId={category?.id} status={status!} />;
+}
+
+export default DemandsPage;
