+++ app/g/[groupSlug]/market/offers/page.tsx
@@ -1,15 +1,30 @@
 'use client';
 
-import { redirect } from 'next/navigation';
 import { groupVar } from '@/lib/apollo/cache';
 import { useReactiveVar } from '@apollo/client';
+import OfferFeed from '@/components/offers/offer-feed';
+import { useDealStatus } from '@/hooks/use-deal-status';
+import { useSearchParams } from 'next/navigation';
 
-function Page() {
-  const group = useReactiveVar(groupVar);
-  if (!group) return <div />;
-
-  const defaultCategorySlug = group?.productCategories[0].slug;
-  return redirect(`offers/${defaultCategorySlug}`);
+export interface OffersPageProps {
+  params: {
+    groupSlug: string;
+  };
 }
 
-export default Page;
+function OffersPage({ params: { groupSlug } }: OffersPageProps) {
+  const group = useReactiveVar(groupVar);
+  const searchParams = useSearchParams();
+  const categorySlug = searchParams.get('category');
+  const category = group?.productCategories.find(
+    (c) => c.slug === categorySlug,
+  );
+  const status = useDealStatus();
+
+  if (!group) return <div />;
+  return (
+    <OfferFeed categoryId={category?.id} status={status!} type="thumbnail" />
+  );
+}
+
+export default OffersPage;
