+++ app/g/[groupSlug]/market/page.tsx
@@ -1,15 +1,9 @@
 'use client';
 
 import { redirect } from 'next/navigation';
-import { groupVar } from '@/lib/apollo/cache';
-import { useReactiveVar } from '@apollo/client';
 
 function Page() {
-  const group = useReactiveVar(groupVar);
-  if (!group) return <div />;
-
-  const defaultCategorySlug = group?.productCategories[0].slug;
-  return redirect(`market/offers/${defaultCategorySlug}`);
+  return redirect(`market/offers`);
 }
 
 export default Page;
