+++ app/g/[groupSlug]/market/swaps/layout.tsx
@@ -11,7 +11,7 @@ interface Props extends SwapsPageProps {
 function SwapsLayout({ params, children }: Props) {
   return (
     <div className="pb-4 md:pb-6">
-      <CategoriesNavbar categorySlug={params.categorySlug} />
+      <CategoriesNavbar />
       <FeedHomeLayout>
         <ThumbnailFeedLayout>{children}</ThumbnailFeedLayout>
       </FeedHomeLayout>
