+++ app/g/[groupSlug]/market/swaps/page.tsx
@@ -1,15 +1,28 @@
 'use client';
 
-import { redirect } from 'next/navigation';
 import { groupVar } from '@/lib/apollo/cache';
 import { useReactiveVar } from '@apollo/client';
+import SwapFeed from '@/components/swaps/swap-feed';
+import { useDealStatus } from '@/hooks/use-deal-status';
+import { useSearchParams } from 'next/navigation';
 
-function Page() {
-  const group = useReactiveVar(groupVar);
-  if (!group) return <div />;
-
-  const defaultCategorySlug = group?.productCategories[0].slug;
-  return redirect(`swaps/${defaultCategorySlug}`);
+export interface SwapsPageProps {
+  params: {
+    groupSlug: string;
+  };
 }
 
-export default Page;
+function SwapsPage({ params: { groupSlug } }: SwapsPageProps) {
+  const group = useReactiveVar(groupVar);
+  const searchParams = useSearchParams();
+  const categorySlug = searchParams.get('category');
+  const category = group?.productCategories.find(
+    (c) => c.slug === categorySlug,
+  );
+  const status = useDealStatus();
+
+  if (!group) return <div />;
+  return <SwapFeed categoryId={category?.id} status={status!} />;
+}
+
+export default SwapsPage;
